HTTP Post with different json format

+++ sensors_logger/cpp_sensors_server/serial.cpp
@@ -452,7 +452,8 @@ void Serial::processLine(NodeMap_t &nodes)
 			temper.time = logbuf.time_now;
 
 			std::string t_temper = notif_map["Temperature"];
-			int l_temper = std::stof(t_temper);
+			//std::cout << "Temperature>" << t_temper
+			float l_temper = std::stof(t_temper);
 			temper.value = l_temper;
 			
 			nodes[l_Id]["Temperature"].push_back(temper);
